avoid unncessary out of place writes
